Use common sources for scala-next and scala3 compiler plugin
